Use direct endpoint entries in default configuration
As done in the examples in the documentation, we originally had an
opinionated template for the endpoints which would only allow for
setting one endpoint to one backend. While this was nice for testing,
this is not ideal if we want to allow folks to have more control over
their endpoints.

So, we now are switching to allowing a raw endpoint configuration
insertion via Krakend's flexible configuration partials [1]. While
service authors will need be more acquainted with lura configuration,
specifically the endpoint constructs [2], it allows us to iterate faster
on verifying the krakend gateway and getting that deployed.

[1] https://www.krakend.io/docs/configuration/flexible-config/
[2] https://www.krakend.io/docs/endpoints/

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>