Enable importing endpoint configuration from a container image
Krakend.io endpoints may take most of the configuration and can get
quite big. Big enough that the ConfigMap's 1.5Mb of storage is not
enough. To address this limitation, we now allow for users to specify if
they want to import endpoints from a container image. This can be
enabled via the `krakend.endpoints.fromImage` value. However, since we
don't know what image they'll use, they need to specify this too.

So, a minimal values file that enables this configuration would look as
follows:

```yaml
krakend:
  krakend:
    endpoints:
        fromImage: true
        image:
          registry: "foo"
          repository: "bar"
          tag: "baz"
```

The idea is that folks will be able to generate images that will be
copied into the main API Gateway container's partials configuration.

Note that it's also possible to override the command to run in order to
copy these files, however, take into account that the resulting
"endpoints" configuration **must** have the name `endpoints.tmpl`.

The default command expects the container to have a file called
`endpoints.json`.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>