Separate port and extra_config settings from default config
The port should ideally be set directly via the already existing
configuration in the helm chart. Doing it outside is prone to errors.
So now the krakend's port is configured via the service's target port,
similarly to how the pod's port is configured.

The extra config is something folks will want to change often, so we
explicitly add a key for people to be able to configure this if they're
using the default configuration we ship. Folks can still override this
and do it on their own if they override the config key.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>