Add support for Argo Rollouts as a deployment
Argo Rollouts is a project that allows us to replace Kubernetes
Deployments and handle more complex deployment strategies [1]. It
currently supports canary and blue green deployments, with the added
value of providing a way of doing analysis on how a rollout is going
based on metrics (e.g. prometheus).

Testing this out, it seems like quite an ideal way of managing the
krakend API Gateway as it would greatly reduce the risk of rolling out
changes if we set up the rollout strategy and analysis templates
appropriately.

Testing this out locally, it seems like quite a versatile tool. However,
I wanted to get some more feedback first.

One caveat I see is that the Kuberentes CRD's are v1alpha1, which is not
ideal.

[1] https://argo-rollouts.readthedocs.io/en/stable/

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>