Back to basics: Use all-in-one image instead of complicated deployment
Instead of enforcing complex deployment strategies such as having a
dedicated endpoints image, we instead allow for an image with all the
configuration backed-in.

This allows for a simpler deployment as we would no longer require
complex initContainers nor specialized images.

This thus removes the usage of the endpoints copier and even the
endpoints image option.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>